Split tracker by transaction type
