+++ src/common/tracker/index.js
@@ -1,3 +1,11 @@
-import TransactionTracker from './TransactionTracker'
+import BackgroundTaskTracker from './BackgroundTaskTracker'
+import RootchainTransactionTracker from './RootchainTransactionTracker'
+import ChildchainTransactionTracker from './ChildchainTransactionTracker'
+import ProcessExitTransactionTracker from './ProcessExitTransactionTracker'
 
-export { TransactionTracker }
+export {
+  BackgroundTaskTracker,
+  RootchainTransactionTracker,
+  ChildchainTransactionTracker,
+  ProcessExitTransactionTracker
+}
